0.0.7
Separate validators and ruls
