+++ lib/proto-mongo.js
@@ -2,6 +2,7 @@
 Object.defineProperty(exports, "__esModule", { value: true });
 const _ = require("lodash");
 const validators_1 = require("./validators");
+const rules_1 = require("./rules");
 exports.rules = {
     types: {
         data: ['undefined', 'string', 'number', 'boolean', 'object', 'array'],
@@ -30,7 +31,7 @@ exports.rules = {
         from: { args: ['!alias'] },
         select: {
             unique: true, all: ['!returns', '!from', '!and', '!orders', '!limit', '!skip'],
-            handle: validators_1.rules.expressions.select.handle,
+            handle: rules_1.rules.expressions.select.handle,
         },
     },
 };
