+++ lib/proto-sql.js
@@ -2,6 +2,7 @@
 Object.defineProperty(exports, "__esModule", { value: true });
 const _ = require("lodash");
 const validators_1 = require("./validators");
+const rules_1 = require("./rules");
 exports.rules = {
     types: {
         data: ['string', 'number', 'boolean'],
@@ -41,7 +42,7 @@ exports.rules = {
         from: { all: ['!alias'] },
         select: {
             unique: true, all: ['!returns', '!from', '!and', '!orders', '!groups', '!limit', '!skip'],
-            handle: validators_1.rules.expressions.select.handle,
+            handle: rules_1.rules.expressions.select.handle,
         },
         union: { all: [':fetch'] },
         unionall: { all: [':fetch'] },
