+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "ancient-babilon",
-  "version": "0.0.6",
+  "version": "0.0.7",
   "description": "One data select language will rule them all.",
   "bugs": "https://github.com/AncientSouls/Babilon/issues",
   "repository": "AncientSouls/Babilon",
